seperated store functions into their own files: raw functionality, error handling, currying, also extracted got reducer
