+++ packages/got-store/package.json
@@ -45,7 +45,7 @@
     },
     "scripts": {
         "test": "jest ./src --passWithNoTests",
-        "test:fn": "jest ./src/tests/$1.spec.js --passWithNoTests",
+        "test:fn": "jest ./src/tests/$0.spec.js --passWithNoTests",
         "build": "webpack --config webpack.config.js; babel src -d dist/cjs --ignore 'src/**/*.spec.js','src/tests'"
     },
     "jest": {
