+++ packages/got-store/src/tests/push.spec.js
@@ -4,7 +4,7 @@ import { Blob } from 'buffer';
 import { toPromise } from '@gothub-team/got-util';
 import { MISSING_PARAM_ERROR } from '../errors.js';
 import { createTestStore } from './shared.js';
-import { GOT_ACTION_UPLOAD_COMPLETE, GOT_ACTION_UPLOAD_PROGRESS } from '../index.js';
+import { GOT_ACTION_UPLOAD_COMPLETE, GOT_ACTION_UPLOAD_PROGRESS } from '../reducer.js';
 
 describe('store:push', () => {
     describe('Basic', () => {
