+++ packages/got-store/src/tests/reducer/mergeOverwrite.spec.js
@@ -1,4 +1,4 @@
-import { GOT_ACTION_MERGE_OVERWRITE, gotReducer } from '../../index.js';
+import { GOT_ACTION_MERGE_OVERWRITE, gotReducer } from '../../reducer.js';
 
 describe('reducer::mergeOverwrite', () => {
     test('should merge nodes into existing graph', () => {
